Fix test with json rounding issues:
This change addresses a possible confusion with JSONS processors
trimming significant zero digits from the fractional part of floating
numbers. It changes zeros to nonzeros in the fractional part, as a way
to enforce the processors to keep the digits.

Another possible solution is to keep the YAML as-is, and ensure the
json shows the prices as 450.00 and 2392.00; ie to change the json
part to be string-equal to the yaml part.